broken: refactor out of ec-core
