+++ Cargo.toml
@@ -1,2 +1,2 @@
 [workspace]
-members = ["constraints", "core"]
+members = ["constraints", "core", "acl", "evm"]
