+++ constraints/src/constraints.rs
@@ -1,36 +1,55 @@
-//! Contains the traits and implementations of each supported constraint.
+use alloc::vec::Vec;
+use core::fmt::Debug;
 
-use ec_core::{Acl, AclKind};
-use ethers_core::types::{NameOrAddress, H160};
+pub use acl::*;
+use codec::MaxEncodedLen;
+use codec::{Decode, Encode};
+use scale_info::TypeInfo;
+use serde::{Deserialize, Serialize};
 
-use crate::{Architecture, Error, Evm};
-
-/// Constraints must implement an evaluation trait that parses.
-pub trait Evaluate<A: Architecture> {
-    fn eval(self, tx: A::TransactionRequest) -> Result<(), Error>;
+/// Supported architectures.
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Encode, Decode, Serialize, Deserialize, TypeInfo)]
+pub enum Arch {
+    Evm,
+    /// Not yet supported on the client, supported in Substrate storage/ACL
+    Btc,
 }
 
-// TODO This can likely be made generic over any architecture with GetRecipient and GetSender traits
-#[allow(clippy::needless_collect)]
-impl Evaluate<Evm> for Acl<[u8; 20]> {
-    fn eval(self, tx: <Evm as Architecture>::TransactionRequest) -> Result<(), Error> {
-        if tx.to.is_none() {
-            return match self.allow_null_recipient {
-                true => Ok(()),
-                false => Err(Error::Evaluation("Null recipients are not allowed.")),
-            };
-        }
+/// Represents a user's constraints
+#[derive(Default, Encode, Decode, Debug, PartialEq, Eq, Clone, TypeInfo)]
+pub struct Constraints {
+    pub evm_acl: Option<Acl<[u8; 20]>>,
+    pub btc_acl: Option<Acl<[u8; 32]>>,
+}
 
-        let converted_addresses: Vec<NameOrAddress> = self
-            .addresses
-            .into_iter()
-            .map(|a| NameOrAddress::Address(H160::from(a)))
-            .collect();
+/// This includes common types and functions related to using ACL functionality.
+mod acl {
+    use super::*;
 
-        match (converted_addresses.contains(&tx.to.unwrap()), self.kind) {
-            (true, AclKind::Allow) => Ok(()),
-            (false, AclKind::Deny) => Ok(()),
-            _ => Err(Error::Evaluation("Transaction not allowed.")),
+    /// An access control list (Allow/Deny lists).
+    #[derive(Clone, Debug, Encode, Decode, PartialEq, Eq, scale_info::TypeInfo, MaxEncodedLen)]
+    pub struct Acl<Address> {
+        pub addresses: Vec<Address>,
+        pub kind: AclKind,
+        pub allow_null_recipient: bool,
+    }
+
+    /// Represents either an allow or deny list.
+    #[derive(Debug, Clone, Copy, PartialEq, Eq, Encode, Decode, TypeInfo, MaxEncodedLen)]
+    pub enum AclKind {
+        Allow,
+        Deny,
+    }
+
+    /// Creates an empty ACL that always evaluates to false.
+    impl<A: Default> Default for Acl<A> {
+        fn default() -> Self {
+            let addresses = Vec::<A>::default();
+            Self {
+                addresses,
+                kind: AclKind::Allow,
+                allow_null_recipient: false,
+            }
         }
     }
 }
