+++ constraints/src/lib.rs
@@ -1,21 +1,8 @@
-use thiserror::Error;
+//! Includes types and interfaces that are foundational to the core of constraints.
 
-#[cfg(test)]
-mod tests;
+#![cfg_attr(not(feature = "std"), no_std)]
 
-pub mod architectures;
-pub mod constraints;
+extern crate alloc;
 
-pub use architectures::*;
 pub use constraints::*;
-
-/// Errors related to parsing and evaulating constraints.
-#[derive(Error, Debug, PartialEq)]
-pub enum Error {
-    /// Transaction request could not be parsed
-    #[error("Invalid transaction request: {0}")]
-    InvalidTransactionRequest(String),
-    /// Transaction request did not meet constraint requirements.
-    #[error("Constraint Evaluation error: {0}")]
-    Evaluation(&'static str),
-}
+pub mod constraints;
