+++ core/Cargo.toml
@@ -3,24 +3,16 @@ name = "ec-core"
 version = "0.1.0"
 edition = "2021"
 
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
 [dependencies]
-codec = { package = "parity-scale-codec", version = "3.0.0", default-features = false }
-scale-info = { version = '2.0.1', default-features = false, features = [
-  'derive',
-] }
-serde = { version = "1.0", default-features = false, features = ["derive"] }
-serde_derive = "1.0.147"
-sp-core = { package = "sp-core", git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.31", default-features = false }
-sp-std = { package = "sp-std", git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.31", default-features = false }
-frame-support = { package = "frame-support", git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.31", default-features = false }
-node-primitives = { version = "2.0.0", default-features = false, git = 'https://github.com/paritytech/substrate.git', branch = "polkadot-v0.9.31" }
+thiserror = "1.0"
+
+# parsing
+serde = { version = "1.0.147", default-features = false, features = ["derive"] }
 
 [features]
 default = ["std"]
 std = [
-  "scale-info/std",
   "serde/std",
-  "frame-support/std",
-  "sp-core/std",
-  "sp-std/std",
 ]
