+++ core/README.md
@@ -1,3 +1,3 @@
 # ec-core
 
-This includes all types and interfaces to be used when using `constraints` related packages.
+This includes core types and interfaces for supporting arbitrary constraints on all signature requests, and architectures for transaction requests.
