+++ evm/src/lib.rs
@@ -1,52 +1,11 @@
-//! This includes the supported architectures and traits for adding new ones.
 
-use ec_core::Arch;
-pub use evm::*;
-use serde::{Deserialize, Serialize};
-
-use crate::Error;
-
-/// Trait for defining important types associated with an architecture.
-pub trait Architecture: Serialize + for<'de> Deserialize<'de> {
-    /// Account type for that chain(SS58, H160, etc)
-    type Address: Eq + Serialize + for<'de> Deserialize<'de>;
-    /// Account type as it is stored in the database
-    type AddressRaw: Eq + Serialize + for<'de> Deserialize<'de>;
-    /// Transaction request type for unsigned transactions
-    type TransactionRequest: GetSender<Self>
-        + GetReceiver<Self>
-        + Serialize
-        + for<'de> Deserialize<'de>
-        + Parse<Self>;
-}
-
-/// Trait for getting the the sender of a transaction.
-pub trait GetSender<A: Architecture> {
-    fn sender(&self) -> Option<A::Address>;
-}
-
-/// Trait for getting the the receiver of a transaction.
-pub trait GetReceiver<A: Architecture> {
-    fn receiver(&self) -> Option<A::Address>;
-}
-
-/// Trait for parsing a raw transaction request into its native transaction request struct.
-pub trait Parse<A: Architecture> {
-    fn parse(raw_tx: String) -> Result<A::TransactionRequest, Error>;
-}
-
-/// Trait for getting the Arch of a transaction request.
-pub trait GetArch {
-    fn arch() -> Arch;
-}
 
 /// EVM architecture
-pub mod evm {
     pub use ethers_core::types::transaction::request::TransactionRequest as EvmTransactionRequest;
     use ethers_core::types::{NameOrAddress, H160};
     use rlp::Rlp;
 
-    use super::*;
+    // use super::*;
 
     #[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
     pub struct Evm;
@@ -77,11 +36,6 @@ pub mod evm {
         }
     }
 
-    impl GetArch for <Evm as Architecture>::TransactionRequest {
-        fn arch() -> Arch {
-            Arch::Evm
-        }
-    }
 
     impl Parse<Evm> for <Evm as Architecture>::TransactionRequest {
         fn parse(
@@ -156,4 +110,3 @@ pub mod evm {
             assert!(unsigned_tx.is_err());
         }
     }
-}
